Uniformity naming buttons, style.css cleaned up
The buttons had differing IDs. Now all buttons are named the same way, with "Button" as the last part of the ID, for readability in views. That naming conformity continues in the views and the controllers.
Most of the things in style.css didn't do anything. Cleaned that up and moved some inline style to the style.css