API - pictures and search 0.1
The search functionality now works. The code still has to be cleaned up a lot in order for it to be readable. The pictures are displayed, in the search, but the formattting doesn't work the way it should.
Data is fetched from the API for the dishDetailsView, but is not yet displayed there.