SidebarView working, model cleaned from old code
The sidebarView now has the intended functionality. It updates with correct prices from the API.
TODO is still the floating numbers, they need to be rounded to integers.
The old code in the model has been removed, for readability. Enough of the new functionality is in place to disregard the old for continued work.